requirements.txt
Django==1.11.16
djangorestframework==3.9.0
Markdown==3.0.1
python-decouple==3.1
pytz==2018.3
requests==2.18.4
six==1.11.0
